+++ src/createLoggerMiddleware.js
@@ -11,7 +11,7 @@ expressWinston.responseWhitelist.push('body');
 const MASKED_HEADERS = ['Cookie', 'cookie', 'Authorization', 'authorization'];
 const MASKED_HEADER_VALUE = '*****';
 
-export default ({logger, MAX_GRAPHQL_VARIABLES_LOG_LENGTH = 512}) => {
+export default ({logger, maxGraphQLVariablesLength = 512}) => {
   const router = new Router();
   return router.use(
     bodyParser.json(),
@@ -34,16 +34,16 @@ export default ({logger, MAX_GRAPHQL_VARIABLES_LOG_LENGTH = 512}) => {
         if (req.method === 'POST' && req.body && req.body.operationName) {
           let variables;
 
-          if (MAX_GRAPHQL_VARIABLES_LOG_LENGTH > 0) {
+          if (maxGraphQLVariablesLength >= 0) {
             const stringifiedVariables = JSON.stringify(req.body.variables);
             const isTooLarge =
-              stringifiedVariables.length > MAX_GRAPHQL_VARIABLES_LOG_LENGTH;
+              stringifiedVariables.length > maxGraphQLVariablesLength;
 
             if (isTooLarge) {
               variables = `${stringifiedVariables.substring(
                 0,
-                MAX_GRAPHQL_VARIABLES_LOG_LENGTH
-              )} […] max payload length reached (${MAX_GRAPHQL_VARIABLES_LOG_LENGTH} chars)`;
+                maxGraphQLVariablesLength
+              )} […] max payload length reached (${maxGraphQLVariablesLength} chars)`;
             } else {
               variables = req.body.variables;
             }